+++ fxn/cli/__init__.py
@@ -9,7 +9,7 @@ from .auth import app as auth_app
 from .env import app as env_app
 from .misc import cli_options
 from .predict import predict
-from .predictors import archive_predictor, create_predictor, delete_predictor, retrieve_predictor, search_predictors
+from .predictors import archive_predictor, create_predictor, delete_predictor, list_predictors, retrieve_predictor, search_predictors
 from ..version import __version__
 
 # Define CLI
@@ -32,6 +32,7 @@ app.add_typer(env_app, name="env", help="Manage predictor environment variables.
 app.command(name="create", help="Create a predictor.")(create_predictor)
 app.command(name="delete", help="Delete a predictor.")(delete_predictor)
 app.command(name="predict", help="Make a prediction.", context_settings={ "allow_extra_args": True, "ignore_unknown_options": True })(predict)
+app.command(name="list", help="List predictors.")(list_predictors)
 app.command(name="search", help="Search predictors.")(search_predictors)
 app.command(name="retrieve", help="Retrieve a predictor.")(retrieve_predictor)
 app.command(name="archive", help="Archive a predictor.")(archive_predictor)